feat: refactor link updating
